Finally optimised X and Y convolutions and they run pretty much around the speed of nvidia's implementation, with filter of 11 running faster than their 17 and filter 23 running not much slower than their 17. Filter 29 currently runs fairly slow at only 600MPixel/sec and is kind of holding the performance back wrt the convolutions, but at least they all work now at a decent speed.
The convolution functions are all now written pretty much the same (excluding the
size 7, which is fine) except for src/dst and filter sizes hard-coded, and define
parameters preceding each function especially. Might(?) leave them that way.

On to tidying up the code and looking at the transpositions again next.