Convolutions 7's are optimised too now, the whole convolution stage as well as the first transposition runs at 55ms for 1024x1024 image (around 18MPixel/sec maybe...)
The final transposition is attempted to be done in blocks of 512x512
now and the kernel has, supposedly, been adjusted for this without having
run it yet. The code in C needs to be written to run these blocks for
all the sigmas and then read everything back if all is working.