Got daisy kernel and C code working!
It currently works at 150ms for 1024x1024 image (i.e. <40ms for 512x512!)

This is without the data transfers. This is done with square blocks of 512x512
and the purpose of this commit is to move away from that to rectangular blocks
of whole lines at a time with an arbitrary number of rows so the data can be
copied back onto a host array and aligned easily. The next commit will hopefully
vouch for this success.