Switch to rectangular blocks in the final transposition appears to have been successful without problems! The operations have been run all together and produced what is meant to be a full DAISY descriptor image and this can be saved as a binary for verification/validation.
The current times for the final outputs are not great they make the whole
thing reach 0.6 seconds solely because of the final data transfer latencies,
there are a few plans to reduce these significantly after the correctness
tests have been done.

The actual numbers seem quite different from the CPU version of DAISY but if
all the operations performed on the GPU pass their correctness tests
then a basic nearest-neighbour correspondence will be attempted on a stereo
pair of images.