All operations done on GPU and on host during the DAISY computations have been checked and verified to be working as intended.
Normalisation was added per set of 8 gradients ('partial') and this was
incorporated along with the gradient transposition kernel since the required
data to do it is available at that stage and therefore the computation can
simply happen along with the transp.

DAISY descriptors were generated for several images in the middlebury
data and compared with the CPU daisy program using matchTest.py and a search
window more than large enough to cover the movement in each pair of images.

The results are x,y offsets and abs diff, compared by plotting them with
plotMatches.py.