Operations 'individually' of the DAISY implementation are considered to be optimised now as well as the transfers back to host have been arranged to overlap with computation** but they are still the bottleneck undoubtedly and not much can be done about that simply due to the bandwidth of any scenario of device-host transfer (i.e. no matter if it is pinned or pageable or DMA or mapped access) the speed of the transfer cannot be reduced with this amount of that. Meaning that, ironically, it is not the latency of memory access that is a problem in this case of memory bottlenecks!
But transferring to pinned host memory is a lot faster than pageable
even with the overhead of allocating it as well as the overhead
of doing the memcpy from the pinned memory onto pageable as a host-host
transfer. Or at least this is the case with multiple DAISY blocks of 200MB,
whereas with only one the case, in terms of the overheads, will be different.

todo: in case of one block don't do the memcpy host-host, the daisy descriptors
      array can remain in the pinned section (presumably..?).

The whole daisy computation has been sped up now by a significant 60ms(!!!)
simply by removing all the printf's and cleaning up other redundant operations
in preparation for the benchmarking! Output of this state of the program
has been generated to be correctness-verified with the monopoly dataset.

The program is now being prepared for the range of inputs it will be tested
on and the recording of the times.