Added DAISY_PROFILING optional code parts and DAISY_HOST_TRANSFERS parts to program, making transfers back to cpu an optional function.
Also the program can now sort of handle non-512x512 multiples of image sizes,
but not non-16 multiples of block heights, this can and will be easily
fixed in the next commit.

Also, the single-block case has been sped up by 20-30ms by not copying
the data from pinned memory to pageable on the host. The data seems to be
usable normally from pinned memory.

todo: - fix the irregular dimension handling in the blocks
      - fix the memory problems of large images on GPU, a limit will have to be introduced by the program
          unless large image handling is implemented (the thought was 1024x1024) but not now.