Got smooth+downsample working for second layer of the pyramid.
Speeds are much much greater as expected.

Now working on third layer (the whole pyramid process will eventually be done
in a loop, the variables are structured for this to be quite straight-forward)
and then will be doing the transposeGradients operation on the pyramid,
might do that in one iteration per layer for ease of development and simplification
of the existing kernel. Normalisation will piggy-back as before.