Transpose gradients works for downsampled data now, and very fast!
Speedup from non-downsampled is about x4.5 for these operations, this is
with a sigma rate of 1 or 1.25. Sigma rate of 1.5 takes it down to x2.9

Going to see if transposeDaisy could easily be adjusted in order to
verify functionality the usual way. Otherwise the search will have to be
implemented and verification to be achieved in that way.